+++ test/__init__.py
@@ -24,8 +24,8 @@ def onlyPy26OrOlder(test):
 
     @functools.wraps(test)
     def wrapper(*args, **kwargs):
-        msg = "{name} requires Python2.7.x+ to run".format(name=test.__name__)
-        if sys.version_info > (2, 6):
+        msg = "{name} only runs on Python2.6.x or older".format(name=test.__name__)
+        if sys.version_info >= (2, 7):
             raise SkipTest(msg)
         return test(*args, **kwargs)
     return wrapper